render game screen
