+++ js/render-screen.js
@@ -1,6 +1,7 @@
 import getIntro from './screen-intro';
 import getGreeting from './screen-greeting';
 import getRuls from './screen-rules';
+import getGame from './game-screen';
 import getGame1 from './screen-game-1';
 import getGame2 from './screen-game-2';
 import getGame3 from './screen-game-3';
@@ -18,9 +19,9 @@ const getScreen = {
   'intro': getIntro,
   'greeting': getGreeting,
   'rules': getRuls,
-  'game1': getGame1,
-  'game2': getGame2,
-  'game3': getGame3,
+  'game1': (state) => getGame(state, getGame1(state)),
+  'game2': (state) => getGame(state, getGame2(state)),
+  'game3': (state) => getGame(state, getGame3(state)),
   'stats': getStats,
 };
